% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionDetector.R
\name{InteractionDetector}
\alias{InteractionDetector}
\title{Interaction Detector}
\description{
Detects important pairwise interactions following the methodology proposed in Lou et al. (2013)
FIXME: cite

Only works with integer or numeric features.
Logical features must be converted to integers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task})\cr
Task.}

\item{\code{xs}}{(\link[data.table:data.table]{data.table::data.table})\cr
data.table of features of the \link[mlr3:Task]{mlr3::Task}.}

\item{\code{n_features}}{(integer(1))\cr
Number of features part of the \link[mlr3:Task]{mlr3::Task}.}

\item{\code{feature_names}}{(character()) \cr
The names of the features of the \link[mlr3:Task]{mlr3::Task}.}

\item{\code{feature_types}}{(character()) \cr
The types of the features of the \link[mlr3:Task]{mlr3::Task}.}

\item{\code{y}}{(numeric())\cr
Numeric target vector of the \link[mlr3:Task]{mlr3::Task}.}

\item{\code{grids}}{(list())\cr
List of grids of size \code{grid_size} containing a grid of each feature.}

\item{\code{rss}}{(matrix())\cr
Symmetric numeric matrix of dimension \code{n_features} times \code{n_features} containing the residual sum of squares for each pair of features.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InteractionDetector-new}{\code{InteractionDetector$new()}}
\item \href{#method-InteractionDetector-compute_best_rss}{\code{InteractionDetector$compute_best_rss()}}
\item \href{#method-InteractionDetector-get_eqcs_from_top_k}{\code{InteractionDetector$get_eqcs_from_top_k()}}
\item \href{#method-InteractionDetector-clone}{\code{InteractionDetector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractionDetector-new"></a>}}
\if{latex}{\out{\hypertarget{method-InteractionDetector-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionDetector$new(task, grid_size = 10L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{([mlr3::Task)\cr
Task.}

\item{\code{grid_size}}{(integer(1))\cr
Grid size used to construct grids for each feature.
Default is \code{10}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractionDetector-compute_best_rss"></a>}}
\if{latex}{\out{\hypertarget{method-InteractionDetector-compute_best_rss}{}}}
\subsection{Method \code{compute_best_rss()}}{
Method to compute the residual sum of squares for each pair of features, overwriting \verb{$rss}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionDetector$compute_best_rss()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible (\code{NULL})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractionDetector-get_eqcs_from_top_k"></a>}}
\if{latex}{\out{\hypertarget{method-InteractionDetector-get_eqcs_from_top_k}{}}}
\subsection{Method \code{get_eqcs_from_top_k()}}{
Retrieves equivalence classes (groups of features) based on the top k most important pairwise interactions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionDetector$get_eqcs_from_top_k(k = 1L, features = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{(integer(1))\cr
The number of top interactions to consider.}

\item{\code{features}}{(character() | NULL)\cr
The features to consider for detecting interactions.
If not provided, all features will be used (default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(A named (integer()) vector indicating the equivalence class each feature belongs to
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractionDetector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InteractionDetector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionDetector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
