% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{Probs}
\alias{Probs}
\title{Probs}
\description{
Class to represent probabilities used in mutation and crossover operations.
}
\details{
This class provides a convenient way to manage and update probabilities used in mutation and crossover operations.
It allows you to set individual probabilities for various operations, such as overall mutation, overall crossover,
parameter mutation, parameter crossover, and group structure mutation.
Future more flexible classes are likely to inherit from this base class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{p_overall_mutate}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the overall mutation probability.}

\item{\code{p_overall_crossover}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the overall crossover probability.}

\item{\code{p_param_mutate}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the hyperparameter mutation probability.}

\item{\code{p_param_crossover}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the hyperparameter crossover probability.}

\item{\code{p_groupstructure_mutate}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the group structure mutation probability.}

\item{\code{p_groupstructure_crossover}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the group structure crossover probability.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Probs-new}{\code{Probs$new()}}
\item \href{#method-Probs-update}{\code{Probs$update()}}
\item \href{#method-Probs-clone}{\code{Probs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Probs-new"></a>}}
\if{latex}{\out{\hypertarget{method-Probs-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Probs$new(
  p_overall_mutate = 0.3,
  p_overall_crossover = 0.7,
  p_param_mutate = 0.2,
  p_param_crossover = 0.5,
  p_groupstructure_mutate = 0.2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p_overall_mutate}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the overall mutation probability.
The default value is 0.3.}

\item{\code{p_overall_crossover}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the overall crossover probability.
This overall crossover probability is also used for the group structure crossover probability.
The default value is 0.7.}

\item{\code{p_param_mutate}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the hyperparameter mutation probability.
The default value is 0.2.}

\item{\code{p_param_crossover}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the hyperparameter crossover probability.
The default value is 0.5.}

\item{\code{p_groupstructure_mutate}}{(numeric(1))\cr
Numeric value between 0 and 1 specifying the group structure mutation probability.
The default value is 0.2.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Probs-update"></a>}}
\if{latex}{\out{\hypertarget{method-Probs-update}{}}}
\subsection{Method \code{update()}}{
Updates the probabilities.

Currently has no effect because probabilities are treated as non-adaptive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Probs$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible (\code{NULL})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Probs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Probs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Probs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
