% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reconstruct_eagga_model}
\alias{reconstruct_eagga_model}
\title{Reconstructs an EAGGA model from a Tuning Instance}
\usage{
reconstruct_eagga_model(instance, tuner, model_uhash)
}
\arguments{
\item{instance}{(\link[mlr3tuning:TuningInstanceMultiCrit]{mlr3tuning::TuningInstanceMultiCrit})\cr
The terminated tuning instance.}

\item{tuner}{(\link{TunerEAGGA})\cr
The tuner that optimized the instance.}

\item{model_uhash}{(\code{character(1)}\cr
A character string representing the unique hash (uhash) of the model to be reconstructed.
See \code{instance$archive$data$uhash} for possible values.}
}
\value{
The reconstructed EAGGA model in the form of the (\link[mlr3:Learner]{mlr3::Learner}) passed during construction of the
\link[mlr3tuning:TuningInstanceMultiCrit]{mlr3tuning::TuningInstanceMultiCrit} trained on the full \link[mlr3:Task]{mlr3::Task}.
}
\description{
This function reconstructs an EAGGA model from a \link[mlr3tuning:TuningInstanceMultiCrit]{mlr3tuning::TuningInstanceMultiCrit} and \link{TunerEAGGA}
and a model unique hash (uhash) value after a completed tuning run
}
\details{
This function reconstructs an EAGGA model observed during tuning specified by its unique hash
by extracting the necessary information from the provided \link[mlr3tuning:TuningInstanceMultiCrit]{mlr3tuning::TuningInstanceMultiCrit} and \link{TunerEAGGA} object.
It clones the learner and task from the tuning instance, sets the learner's hyperparameter values according
to the values logged into the archive matching the given unique hash of the model and trains the learner on the task.
Note that if no random seed was specified during construction of the \link{TunerEAGGA} (via the
\code{seed_calculate_proxy_measures} parameter), exact reconstruction may not be possible.
As the learner is trained on the complete task, the model can be readily used on new data.
It should no longer be used for performance estimation unless you do have access to another task containing hold-out
test data that was not seen during tuning.
}
